pyproject.toml
[tool.poetry]
name = "animal-shelter-django"
version = "0.1.0"
description = "App for animal shelters"
authors = ["Mateusz Pydych <mpydych@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
black = "^22.6.0"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
model-bakery = "^1.7.0"
django-extensions = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
